Cleaned warnings in db.cpp. Removed function-like macros.

+++ src/entities/char_data.h
@@ -367,7 +367,7 @@ class CharData : public ProtectedCharData {
 	bool HaveFeat(EFeat feat_id) const { return real_abils.Feats.test(to_underlying(feat_id)); };
 
 	void set_skill(const ESkill skill_id, int percent);
-	void SetSkillAfterRemort(short remort);
+	void SetSkillAfterRemort(int remort);
 	void clear_skills();
 	int GetSkill(const ESkill skill_id) const;
 	int GetSkillWithoutEquip(const ESkill skill_id) const;
